land: retype pass 3 — 485 landed functions on PDB-named engine headers (byte-exact)
- gen_class_headers.py flattens leaf-only wrappers (CCountedPointer<T> -> Name_Data/Name_Info,
  scoped_ptr, CDefPointer), fills unlisted/inherited gaps, drops union alternatives, compile-checks
  every header individually (962 OK), pins headers that retyped files already include, and keeps
  retail-only carves sticky with their original types across regenerations
- retype_landed.py: shim inserted at a marker (no stale index), INLINE_METHODS / INHERITANCE skips
- all 485 retyped TUs compile; dashboard + README regenerated

+++ rebuild/include/engine/CAnimComponentCombatHandedness.h
@@ -11,7 +11,7 @@
 #pragma pack(push, 1)
 struct CAnimComponentCombatHandedness {
     void*         __vftable;                                 // +0x00 vptr, or first dword of a flattened base subobject
-    int           f4;                                        // +0x04 retail-only (CAnimComponentCombatHandedness_CloneCAnimComponentCombatHandednessUBEPAVCAnimCo_005db260.cpp)
+    long          f4;                                        // +0x04 retail-only (previous header)
     unsigned char StartHandedness[0x4];                      // +0x08 ECombatAnimationHandedness
     unsigned char EndHandedness[0x4];                        // +0x0c ECombatAnimationHandedness
 };
